Langchain/requirements.txt
langchain
openai==0.28.1
huggingface_hub
azure-search-documents
torch
azure-identity
python-dotenv
streamlit